// src/types/slideTypes.tsx
// src/types/slideTypes.ts

export type SlideType = {
    id: string;
    image: string;
    title: string;
    description: string;
    details: string;
    date: string;
    time: string;
    startDate: string;
    endDate: string;
    startTime: string;
    endTime: string;
    location: string;
    host: string;
    availableSlots: number;
    totalSlots: number;
    isCreator: boolean;
  };
  